refactor(tasks): split tasks.py god module into domain-focused task modules
Splits backend/app/tasks.py (1,678 LOC, 4 domains) into:
- tasks/sync.py      — 8 data-sync tasks (tickers, aggregates, splits, news, futures, tweet-monitor)
- tasks/scanning.py  — 4 scanner tasks (universe scan, range scan, liquidity hunt, alert evaluation)
- tasks/trading.py   — 3 auto-trade tasks + 5 private fill helpers
- tasks/quality.py   — 3 quality/analysis tasks

tasks/__init__.py re-exports all 18 public task objects so all existing
import paths (from app.tasks import ...) resolve unchanged. Every task
decorator carries name='app.tasks.<task_name>' to preserve Celery string
identity — beat schedule and send_task call sites are untouched.

Test patch paths updated: app.tasks.* → app.tasks.trading.*
New smoke test: tests/tasks/test_package_exports.py (7 assertions)
